requirements.txt
# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0
tensorboardX>=2.6.0

# Jupyter/Interactive Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
gpustat>=1.1.0

# AWS S3 Support (for checkpoint syncing)
boto3>=1.28.0
awscli>=1.29.0

# Development Tools (optional)
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.4.0

